posts: add url for authors (#29)
I initially added an `authorUrl` field to some posts and changed the template to use it, but then I remembered that we already parse everything in metadata as markdown!

So this can be as easy as correctly formatting the author field. Note, however, that the `yaml` parser will freak out about some markdown syntax, so we'll have to quote fields like this.

Reviewed-on: https://codeberg.org/group-meowing/website/pulls/29
Co-authored-by: Evie Ciobanu <me@evie.ro>
Co-committed-by: Evie Ciobanu <me@evie.ro>